requirements.txt
pandas
signs
openai
bophono
botok
pyewts
pdf2image
pyperclip
google-api-python-client
google-auth-httplib2
google-auth-oauthlib